% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulating.R
\name{add_geoms}
\alias{add_geoms}
\title{Add leaves or flowers to QSM}
\usage{
add_geoms(
  qsm,
  geom_distribution,
  geom,
  stem_len = 0.01,
  geom_type = c("leaf", "flower"),
  geom_angle = 45,
  cylinder_classes = NULL,
  add_noise = 0
)
}
\arguments{
\item{qsm}{An object of class \code{QSM}.}

\item{geom_distribution}{\code{data.frame}, describes the geom distribution,
with columns \code{c("diam_start_m", "diam_end_m", "m_per_geom",
"geom_scaling")}, optionally including variables from \code{classify_crown()}.}

\item{geom}{\code{matrix}, contains matrix with IDs and coordinates.}

\item{stem_len}{\code{numeric}, geom stem length.}

\item{geom_type}{\code{character}, type of added geoms, either
\code{"leaves"} or \code{"flowers"}}

\item{geom_angle}{\code{numeric}, the angle between the ground and the leaf
axis in degrees when simulating leaves.}

\item{cylinder_classes}{\code{matrix} (optional), crown classification of
each cylinder when using an geom distribution including variables from
\code{classify_crown()}.}

\item{add_noise}{\code{numeric} (optional), fraction of the stem length used
for adding noise to the geom location and stem length, 0 means no noise.}
}
\value{
\code{matrix}, contains coordinates of the simulated geoms. For each polygon
node of the initial \code{geom}, three columns for the xyz-coordinates
are given.
}
\description{
\code{add_geoms} simulates leaves or flowers for a \code{QSM} object.
}
\examples{
# load qsm
file_path <- system.file("extdata", "walnut.mat", package="qsm2shade")
qsm <- qsm2r::readQSM(file_path)

# shift qsm to origin
# (shade is always projected to z = 0)
qsm <- qsm2r::set_location(qsm, c(0,0,0))

# create polygons for single geom
flower <- create_flower()

# without crown classification:

# create dummy geom regression
distribution <- dummy_geom_distribution()

# add geoms
flowers <- add_geoms(qsm, distribution, flower, geom_type = "flower")

# display values
head(flowers)

# with crown classification:

# classify crown cylinders
classes <- classify_crown(
  qsm, compass_directions = 8, outside_buffer_m = 2, vertical_sections = 3)

# create dummy geom regression
distribution_classes <- dummy_geom_distribution(cylinder_classes = TRUE)

# add geoms
flowers <- add_geoms(
  qsm, distribution_classes, flower, geom_type = "flower",
  cylinder_classes = classes)

# display values
head(flowers)
}
\seealso{
\code{\link{dummy_geom_distribution}},
\code{\link{create_flower}}, \code{\link{create_leaf}}
}
