% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulating.R
\name{add_items}
\alias{add_items}
\title{Add leaves or flowers}
\usage{
add_items(
  qsm,
  item_distribution,
  item_poly,
  stem_len = 0.01,
  item_type = c("leaves", "flowers"),
  item_angle = 45,
  cylinder_classes = NULL,
  add_noise = 0
)
}
\arguments{
\item{qsm}{An object of class \code{QSM}.}

\item{item_distribution}{a}

\item{item_poly}{a}

\item{stem_len}{a}

\item{item_type}{a}

\item{item_angle}{a}

\item{cylinder_classes}{a (optional)}

\item{add_noise}{a (optional)}
}
\value{
A matrix. Contains coordinates of the simulated items. For each polygon node
of the initial \code{item_poly}, three columns for the xyz-coordinates are
given.
}
\description{
\code{add_items} simulates leaves or flowers for a \code{QSM} object.
}
\examples{
# load qsm
file_path <- system.file("extdata", "Prunus_avium_QSM_simplified.mat", package="qsm2shade")
qsm <- qsm2r::readQSM(file_path)

# create polygons for single item
flower <- create_single_flower()

# without classification:

# create dummy item regression
distribution <- dummy_item_distribution()

# add items
add_items(qsm, distribution, flower, item_type = "flowers")

# with classification:

# classify crown cylinders
classes <- classify_crown(qsm)

# create dummy item regression
distribution_classes <- dummy_item_distribution(classification = TRUE)

# add items
add_items(qsm, distribution_classes, flower, item_type = "flowers", cylinder_classes = classes)

}
\seealso{
\code{\link{classify_crown}}, \code{\link{classify_crown}}
}
