% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shading.R
\name{add_radiation}
\alias{add_radiation}
\title{Add radiation values to simulated shade}
\usage{
add_radiation(raster, radiation)
}
\arguments{
\item{raster}{\code{SpatRaster}, stacked set of shade rasters from
\code{shade_tree_qsm()} or \code{shade_tree_geoms()}.}

\item{radiation}{\code{data.frame}, diffuse and global radiation over time,
must be formatted as the example from \code{dummy_radiation()}.}
}
\value{
\code{SpatRaster}, contains radiation for each time period, the
layer names indicate the respective time period.
}
\description{
\code{add_radiation} adds radiation values to the rasters produced by
\code{shade_tree_qsm()} or \code{shade_tree_geoms()}.
}
\details{
The spatial resolution of \code{empty_grid} should correspond
to the unit used in \code{radiation}, e.g. when using a resolution of 0.1,
the radiation should be given per dm². If the raster cell contains any
shade, the diffuse radiation is assigned. Otherwise, the global radiation is
assigned. Turning parallel processing on (\code{sequential = FALSE}) might
only work for windows computers.
}
\examples{
# load wood geoms
file_path <- system.file("extdata", "pear_wood.txt", package="qsm2shade")
geom_wood <- as.matrix(read.table(file_path, header = TRUE))

# load leaf geoms
file_path <- system.file("extdata", "pear_leaves.txt", package="qsm2shade")
geom_other <- as.matrix(read.table(file_path, header = TRUE))

# get sun position at different times
timeframe <- seq(
  ISOdate(2020, 03, 22, 0, 0), ISOdate(2020, 03, 22, 23, 50), "10 mins")
sun_position <- sun_movement(timeframe, latitude = 48.07, longitude = 7.60)

# calculate shade
result <- shade_tree_geoms(
  geom_wood, sun_position = sun_position, geom_other = geom_other,
  transparency = 0.7)

# create dummy radiation data
radiation <- dummy_radiation(
  ISOdate(2020, 01, 01, 0, 0), ISOdate(2020, 12, 31, 23, 50), "1 hour")

# add radiation data to shade polygons
result <- add_radiation(result, radiation)

# show results
print(result)
}
\seealso{
\code{\link{shade_tree_qsm}}, \code{\link{shade_tree_geoms}}, \code{\link{shade_summarize}}
}
