% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifying.R
\name{classify_crown}
\alias{classify_crown}
\title{Classify crown cylinders into crown sections}
\usage{
classify_crown(
  qsm,
  compass_directions = c(0, 4, 8),
  outside_buffer_m = 2,
  vertical_sections = c(0, 2, 3),
  min_cylinders_cbh = 3
)
}
\arguments{
\item{qsm}{An object of class \code{QSM}.}

\item{compass_directions}{\code{integer}, number of compass directions (0,4,8).}

\item{outside_buffer_m}{\code{numeric}, width of the buffer defining the outer and inner
horizontal section in meters.}

\item{vertical_sections}{\code{integer}, number of vertical crown sections.}

\item{min_cylinders_cbh}{\code{integer}, minimum number of cylinders of the lowest branch
defining crown base height.}
}
\value{
\code{matrix}, contains three columns: \code{compass} (compass directions),
\code{vertical} (vertical crown sections, 1 = lowest, n = highest) and
\code{horizontal} (horizontal crown sections).
}
\description{
\code{classify_crown} classifies the cylinders of a \code{QSM} object into
separate  crown sections. It can use a variable number of compass directions
and vertical sections and classify the crown into an outer and an inner
horizontal section.
}
\examples{
# load qsm
file_path <- system.file("extdata", "walnut.mat", package="qsm2shade")
qsm <- qsm2r::readQSM(file_path)

# classify crown cylinders
crown <- classify_crown(qsm, compass_directions = 8, outside_buffer_m = 0.5, vertical_sections = 3)

# add classification to qsm
qsm@cylinder$vertical <- crown[, "vertical"]
qsm@cylinder$vertical[is.na(qsm@cylinder$vertical)] <- "stem"

# plot qsm according to classification
qsm2r::plot(qsm, col_var = "vertical")

# add classification to qsm
qsm@cylinder$horizontal <- crown[, "horizontal"]
qsm@cylinder$horizontal[is.na(qsm@cylinder$horizontal)] <- "stem"

# plot qsm according to classification
qsm2r::plot(qsm, col_var = "horizontal")
}
\seealso{
\code{\link{add_geoms}}
}
