% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulating.R
\name{create_single_flower}
\alias{create_single_flower}
\title{Create single flower polygons}
\usage{
create_single_flower(radius_m = 0.01)
}
\arguments{
\item{radius_m}{\code{numeric}, radius of the pentagon for the flower base
in meters.}
}
\value{
\code{list}, each item contains a \code{data.frame} with coordinates of one
of the polygons of a single flower.
}
\description{
\code{create_single_flower} creates polygons of a single flower to be used
by \code{add_item()}.
}
\examples{
# create polygons for single flower
flower <- create_single_flower()
}
\seealso{
\code{\link{add_item}}, \code{\link{plot_single_item}}
}
