% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulating.R
\name{create_single_leaf}
\alias{create_single_leaf}
\title{Create single leaf polygon}
\usage{
create_single_leaf(
  leaf_type = c("normal", "heart", "feather"),
  length_m = 0.01
)
}
\arguments{
\item{leaf_type}{\code{character}, shape of the leaf, should be one of
\code{c("normal", "heart", "feather")}.}

\item{length_m}{\code{numeric}, length of a single leaf in meters.}
}
\value{
\code{data.frame}, contains coordinates of a single leaf.
}
\description{
\code{create_single_leaf} creates a polygon of a single leaf to be used
by \code{add_item()}.
}
\examples{
# create polygon for single leaf
leaf <- create_single_leaf()
}
\seealso{
\code{\link{add_item}}, \code{\link{plot_single_item}}
}
