% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulating.R
\name{dummy_geom_distribution}
\alias{dummy_geom_distribution}
\title{Create a dummy geom distribution}
\usage{
dummy_geom_distribution(cylinder_classes = FALSE)
}
\arguments{
\item{cylinder_classes}{\code{boolean}, whether the variables from
\code{classify_crown()} should be included.}
}
\value{
\code{data.frame}, contains an geom distribution.
}
\description{
\code{dummy_geom_distribution} creates a dummy geom distribution to be used
by \code{add_geoms()}. When setting \code{classification = TRUE}, 8 compass
directions, 2 horizontal sections and 3 vertical sections are used.
}
\examples{
# create dummy geom regression
distribution <- dummy_geom_distribution()

# display values
print(distribution)
}
\seealso{
\code{\link{add_geoms}}
}
