% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulating.R
\name{dummy_item_distribution}
\alias{dummy_item_distribution}
\title{Create a dummy item distribution}
\usage{
dummy_item_distribution(classification = FALSE)
}
\arguments{
\item{classification}{\code{boolean}, whether the variables from
\code{classify_crown()} should be included.}
}
\value{
\code{data.frame}, contains an item distribution.
}
\description{
\code{dummy_item_distribution} creates a dummy item distribution to be used
by \code{add_item()}. When setting \code{classification = TRUE}, 8 compass
directions, 2 horizontal sections and 3 vertical sections are used.
}
\examples{
# create dummy item regression
distribution <- dummy_item_distribution()
}
\seealso{
\code{\link{add_item}}
}
