% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shading.R
\name{dummy_radiation}
\alias{dummy_radiation}
\title{Create dummy radiation data}
\usage{
dummy_radiation(
  start = ISOdate(2020, 1, 1, 0, 0),
  end = ISOdate(2020, 12, 31, 23, 50),
  interval = "1 hour"
)
}
\arguments{
\item{start}{\code{POSIXct}, start time of the radiation data.}

\item{end}{\code{POSIXct}, end time of the radiation data.}

\item{interval}{\code{character}, time interval.}
}
\value{
\code{data.frame}, contains diffuse and global radiation over time.
}
\description{
\code{dummy_radiation} creates dummy radiation data to be used
by \code{add_radiation()}.
}
\examples{
# create dummy radiation data
radiation <- dummy_radiation(ISOdate(2020, 01, 01, 0, 0), ISOdate(2020, 12, 31, 23, 50), "1 hour")

# display values
head(radiation)
}
\seealso{
\code{\link{add_radiation}}
}
