% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shading.R
\name{dummy_radiation_hourly}
\alias{dummy_radiation_hourly}
\title{Create dummy radiation data}
\usage{
dummy_radiation_hourly(
  start = ISOdate(2020, 1, 1, 0, 0),
  end = ISOdate(2020, 12, 31, 23, 50)
)
}
\arguments{
\item{start}{\code{POSIXct}, start time of the radiation data.}

\item{end}{\code{POSIXct}, end time of the radiation data.}
}
\value{
\code{data.frame}, contains diffuse and global radiation over time.
}
\description{
\code{dummy_radiation_hourly} creates dummy radiation data to be used
by \code{shade_tree()}.
}
\examples{
# create dummy radiation data
radiation_hourly <- dummy_radiation_hourly(ISOdate(2020, 01, 01, 0, 0), ISOdate(2020, 12, 31, 23, 50))
}
\seealso{
\code{\link{shade_tree}}
}
