% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geom_shift}
\alias{geom_shift}
\title{Move geoms by an offset}
\usage{
geom_shift(geom, offset = c(0, 0, 0))
}
\arguments{
\item{geom}{\code{matrix}, contains matrix with IDs and coordinates.}

\item{offset}{\code{numeric}, contains \code{xyz}-coordinates of the require shift.}
}
\value{
A matrix with the IDs and coordinates of the shifted geoms.
}
\description{
\code{geom_shift} shifts the coordinates of geoms by specified offsets in
the x, y and z direction.
}
\examples{
# load wood geoms
file_path <- system.file("extdata", "pear_wood.txt", package="qsm2shade")
geom_wood <- read.table(file_path, header = TRUE)

# get stem location
location <- geom_tree_location(geom_wood)

# display values
summary(geom_wood)

# shift stem location to 0,0,0
geom_wood <- geom_shift(geom_wood, location)

# display values
summary(geom_wood)
}
\seealso{
\code{\link{geom_tree_location}}
}
