% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{las_tree_location}
\alias{las_tree_location}
\title{Extract stem coordinates from a point cloud of a single tree}
\usage{
las_tree_location(las, lwr_height = 1.2, upr_height = 1.4)
}
\arguments{
\item{las}{An object of class \code{LAS}.}

\item{lwr_height}{\code{numeric}, lower height threshold in meters.}

\item{upr_height}{\code{numeric}, upper height threshold in meters.}
}
\value{
A \code{numeric} containing the \code{xyz}-coordinates of the stem center.
}
\description{
\code{las_tree_location} extracts the coordinates of the stem of a single
tree at a specified height range. Assumes that the input \code{LAS} object
contains only one tree. The heights refer to the height over the ground.
}
\examples{
# load las data
file_path <- system.file("extdata", "walnut_with_ground.las", package="qsm2shade")
las <- lidR::readLAS(file_path)

# get stem location
location <- las_tree_location(las)

# display values
print(location)
}
\seealso{
\code{\link{las_tree_location}}, \code{\link{plot_ground}}
}
