% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_geom}
\alias{plot_geom}
\title{Plot a single geom}
\usage{
plot_geom(geom, col = "#6D9DC5", lit = TRUE, axes = TRUE)
}
\arguments{
\item{geom}{\code{matrix}, contains matrix with IDs and coordinates.}

\item{col}{\code{character}, color of the geom.}

\item{lit}{\code{boolean}, whether the polygons should be lit.}

\item{axes}{\code{boolean}, whether to add axes.}
}
\value{
\code{rgl} plot of one geom.
}
\description{
\code{plot_geom} plots a single geom to be used by \code{add_geoms()}.
}
\examples{
# create single flower
flower <- create_flower()

# plot single flower
plot_geom (flower, "pink2")

# or:

# create single leaf
leaf <- create_leaf(type = "normal")

# plot single leaf
plot_geom (leaf, "darkolivegreen3")
}
\seealso{
\code{\link{add_geoms}}, \code{\link{create_flower}},
\code{\link{create_leaf}},
}
