% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_items}
\alias{plot_items}
\title{Plot simulated items}
\usage{
plot_items(item_pts, col, add = TRUE)
}
\arguments{
\item{item_pts}{\code{matrix}, contains coordinates of simulated items.}

\item{col}{\code{character}, color of the items.}

\item{add}{\code{boolean}, add the plot to current active \code{rgl} plot.}
}
\value{
\code{rgl} plot of all items.
}
\description{
\code{plot_items} plots items simulated by \code{add_item()}.
}
\examples{
# load qsm
file_path <- system.file("extdata", "Prunus_avium_QSM_simplified.mat", package="qsm2shade")
qsm <- qsm2r::readQSM(file_path)

# create dummy item regression
distribution <- dummy_item_distribution()

# create polygons for single item
flower <- create_single_flower()

# add items
flowers <- add_items(qsm, distribution, flower, item_type = "flowers")

# plot qsm
qsm2r::plot(qsm, col = "salmon4", lit = TRUE)

# plot items
plot_items(flowers, col = "pink2")

# or:

# create polygons for single item
leaf <- create_single_leaf(leaf_type = "normal", length_m = 0.1)

# add items
leaves <- add_items(qsm, distribution, leaf, item_type = "leaves")

# plot qsm
qsm2r::plot(qsm, col = "salmon4", lit = TRUE)

# plot items
plot_items(leaves, col = "darkolivegreen3")
}
\seealso{
\code{\link{add_item}}, \code{\link{plot_shade_items}}
}
