% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_polys}
\alias{plot_polys}
\title{Plot polygons}
\usage{
plot_polys(poly_geom, col = "grey40", add = TRUE, lit = TRUE)
}
\arguments{
\item{poly_geom}{\code{matrix}, contains matrix with IDs and coordinates.}

\item{col}{\code{character}, color of the items.}

\item{add}{\code{boolean}, add the plot to current active \code{rgl} plot.}

\item{lit}{\code{boolean}, whether the polygons should be lit.}
}
\value{
\code{rgl} plot of all items.
}
\description{
\code{plot_polys} plots items simulated by \code{add_item()}.
}
\examples{
# load wood polygons
file_path <- system.file("extdata", "pear_wood.txt", package="qsm2shade")
poly_wood <- read.table(file_path, header = T)

# load leaf polygons
file_path <- system.file("extdata", "pear_leaves.txt", package="qsm2shade")
poly_leaves <- read.table(file_path, header = T)

# plot wood & leaves (takes quite some time)
plot_polys(poly_wood, col = "salmon4", add = F)
plot_polys(poly_leaves, col = "darkolivegreen3", add = T)
}
\seealso{
\code{\link{add_item}}, \code{\link{plot_shade_items}}
}
