% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_shade_geoms}
\alias{plot_shade_geoms}
\title{Plot shadows cast by geoms}
\usage{
plot_shade_geoms(
  geoms,
  sun_direction = c(0.25, 0.5, -0.75),
  plane_origin = c(0, 0, 0),
  plane_normal = c(0, 0, 1),
  col = "grey40",
  add = TRUE
)
}
\arguments{
\item{geoms}{\code{matrix}, contains coordinates of simulated leaves /
flowers.}

\item{sun_direction}{\code{numeric}, vector containing the unit vector of the
sun direction.}

\item{plane_origin}{\code{numeric}, \code{xyz}-vector of a point on the
ground.}

\item{plane_normal}{\code{numeric}, \code{xyz}-vector of the ground normal.}

\item{col}{\code{character}, color of the shade.}

\item{add}{\code{boolean}, add the plot to current active \code{rgl} plot.}
}
\value{
\code{rgl} plot of the shade.
}
\description{
\code{plot_shade_geoms} plots shadows of geoms simulated by
\code{add_geoms()}. The ground can be specified via a point on the plane and
the plane normal. Per default, an even ground at the coordinate origin is
assumed.
}
\examples{
# load qsm
file_path <- system.file("extdata", "walnut.mat", package="qsm2shade")
qsm <- qsm2r::readQSM(file_path)

# shift qsm to origin
# (shade is always projected to z = 0)
qsm <- qsm2r::set_location(qsm, c(0,0,0))

# create dummy geom regression
distribution <- dummy_geom_distribution()

# create polygons for single geom
leaf <- create_leaf(type = "normal", length_m = 0.1)

# add geoms
leaves <- add_geoms(qsm, distribution, leaf, geom_type = "leaf")

# plot qsm & leaves
qsm2r::plot(qsm, col = "salmon4", lit = TRUE)
plot_geoms(leaves, col = "darkolivegreen3")

# plot shadows
plot_shade_qsm(qsm)
plot_shade_geoms(leaves)
}
\seealso{
\code{\link{add_geoms}}, \code{\link{plot_shade_qsm}}
}
