% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_shade_items}
\alias{plot_shade_items}
\title{Plot shadows cast by items}
\usage{
plot_shade_items(
  item_pts,
  sun_direction = c(0.25, 0.5, -0.75),
  plane_origin = c(0, 0, 0),
  plane_norm = c(0, 0, 1),
  col = "grey40",
  add = TRUE
)
}
\arguments{
\item{item_pts}{\code{matrix}, contains coordinates of simulated items.}

\item{sun_direction}{\code{numeric}, vector containing the unit vector of the
sun direction.}

\item{plane_origin}{\code{numeric}, \code{xyz}-vector of a point on the
ground.}

\item{plane_norm}{\code{numeric}, \code{xyz}-vector of the ground normal.}

\item{col}{\code{character}, color of the shade.}

\item{add}{\code{boolean}, add the plot to current active \code{rgl} plot.}
}
\value{
\code{rgl} plot of the shade.
}
\description{
\code{plot_shade_items} plots shadows of items simulated by
\code{add_item()}. The ground can be specified via a point on the plane and
the plane normal. Per default, an even ground at the coordinate origin is
assumed.
}
\examples{
# load qsm
file_path <- system.file("extdata", "Prunus_avium_QSM_simplified.mat", package="qsm2shade")
qsm <- qsm2r::readQSM(file_path)

# shift qsm to origin
# (shade is always projected to z = 0)
qsm <- qsm2r::set_location(qsm, c(0,0,0))

# create dummy item regression
distribution <- dummy_item_distribution()

# create polygons for single item
flower <- create_single_flower()

# create dummy item regression
distribution <- dummy_item_distribution()

# add items
flowers <- add_items(qsm, distribution, flower, item_type = "flowers")

# plot qsm
qsm2r::plot(qsm, col = "salmon4", lit = TRUE)

# plot items
plot_items(flowers, col = "pink2")

# plot shade of wood
plot_shade_wood(qsm)

# plot shade of items
plot_shade_items(flowers)

# or:

# create polygons for single item
leaf <- create_single_leaf(leaf_type = "normal", length_m = 0.1)

# add items
leaves <- add_items(qsm, distribution, leaf, item_type = "leaves")

# plot qsm
qsm2r::plot(qsm, col = "salmon4", lit = TRUE)

# plot items
plot_items(leaves, col = "darkolivegreen3")

# plot shade of wood
plot_shade_wood(qsm)

# plot shade of items
plot_shade_items(leaves)
}
\seealso{
\code{\link{add_item}}, \code{\link{plot_shade}}
}
