% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_shade_wood}
\alias{plot_shade_wood}
\title{Plot shadows cast by wood}
\usage{
plot_shade_wood(
  qsm,
  sun_direction = c(0.25, 0.5, -0.75),
  col = "grey",
  add = TRUE
)
}
\arguments{
\item{qsm}{An object of class \code{QSM}.}

\item{sun_direction}{\code{numeric}, vector containing the unit vector of the
sun direction.}

\item{col}{\code{character}, color of the shade}

\item{add}{\code{boolean}, add the plot to current active \code{rgl} plot.}
}
\value{
\code{rgl} plot of the shade.
}
\description{
\code{plot_shade_wood} plots shadows of a \code{QSM} object.
}
\examples{
# load qsm
file_path <- system.file("extdata", "Prunus_avium_QSM_simplified.mat", package="qsm2shade")
qsm <- qsm2r::readQSM(file_path)

# plot qsm
qsm2r::plot(qsm, col = "salmon4", lit = TRUE)

# plot shade of wood
plot_shade_wood(qsm)
}
\seealso{
\code{\link{plot_shade_items}}
}
