% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_single_item}
\alias{plot_single_item}
\title{Plot a single item}
\usage{
plot_single_item(item_poly, col = "#F4ACB7")
}
\arguments{
\item{item_poly}{\code{matrix} or \code{list}, item to be simulated,
contains coordinates of a single polygon in a matrix or a list of matrices.}

\item{col}{\code{character}, color of the item.}
}
\value{
\code{rgl} plot of one item.
}
\description{
\code{plot_single_item} plots a single item to be used by \code{add_item()}.
}
\examples{
# create single flower
flower <- create_single_flower()

# plot single flower
plot_single_item (flower)

# or:

#' # create single leaf
leaf <- create_single_leaf(leaf_type = "normal")

# plot single leaf
plot_single_item (leaf)
}
\seealso{
\code{\link{add_item}}, \code{\link{create_single_flower}},
\code{\link{create_single_leaf}},
}
