% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{poly_shift}
\alias{poly_shift}
\title{Move polygons to coordinates}
\usage{
poly_shift(poly, offset = c(0, 0, 0))
}
\arguments{
\item{poly}{contains matrix with IDs and coordinates of polygons.}

\item{offset}{\code{numeric}, contains \code{xyz}-coordinates of the require shift.}
}
\value{
A matrix with the IDs and coordinates of the shifted polygons.
}
\description{
\code{poly_shift} shifts the coordinates of polygons by specified offsets in
the x, y and z direction.
}
\examples{
# load wood polygons
file_path <- system.file("extdata", "pear_wood.txt", package="qsm2shade")
poly_wood <- read.table(file_path, header = T

# get stem location
location <- las_tree_location(poly_wood)

# shift stem location to 0,0,0
poly_wood <- poly_shift(poly_wood, location)
}
\seealso{
\code{\link{poly_tree_location}}
}
