% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shading.R
\name{shade_merge}
\alias{shade_merge}
\title{Merge shade of multiple trees}
\usage{
shade_merge(
  rasters,
  resolution = unique(terra::res(rasters[[1]])),
  background = 1
)
}
\arguments{
\item{rasters}{\code{list}, single or summarised rasters from
\code{shade_tree_qsm()} or \code{shade_tree_geoms()}.}

\item{resolution}{\code{numeric}, spatial resolution of input rasters.}

\item{background}{\code{numeric}, desired raster value of regions covered by
none of the input rasters, shade is coded as 0 and light is coded as 1.}
}
\value{
\code{SpatRaster}, contains light / shade / in-between for each time period,
the layer names indicate the respective time period.
}
\description{
\code{shade_merge} merges rasters produced by
\code{shade_tree_qsm()} or \code{shade_tree_geoms()}, but works only for a
single point in time.
}
\examples{
# load wood geoms
file_path <- system.file("extdata", "pear_wood.txt", package="qsm2shade")
geom_wood <- as.matrix(read.table(file_path, header = TRUE))

# load leaf geoms
file_path <- system.file("extdata", "pear_leaves.txt", package="qsm2shade")
geom_other <- as.matrix(read.table(file_path, header = TRUE))

# get sun position at different times
timeframe <- seq(ISOdate(2020, 03, 22, 10, 0), ISOdate(2020, 03, 22, 10, 0), "10 mins")
sun_position <- sun_movement(timeframe, latitude = 48.07, longitude = 7.60)

# calculate shade
result <- shade_tree_geoms(
  geom_wood, sun_position = sun_position, geom_other = geom_other,
  transparency = 0.7)

# move raster to simulate more trees
result_add_1 <- terra::shift(result, dx = 14, dy = 4)
result_add_2 <- terra::shift(result, dx = 7, dy = -1)
result_add_3 <- terra::shift(result, dx = -8, dy = -3)

# merge rasters
result_merged <- shade_merge(
  list(result, result_add_1, result_add_2, result_add_3))

# show results
terra::plot(result_merged)

# create dummy radiation data
radiation <- dummy_radiation(
  ISOdate(2020, 01, 01, 0, 0), ISOdate(2020, 12, 31, 23, 50), "1 hour")

# add radiation data
result_merged <- add_radiation(result_merged, radiation)

# show results
terra::plot(result_merged)
}
\seealso{
\code{\link{shade_tree_qsm}}, \code{\link{shade_tree_geoms}}, \code{\link{add_radiation}}
}
