% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shading.R
\name{shade_tree}
\alias{shade_tree}
\title{Calculate shade of a QSM}
\usage{
shade_tree(
  qsm,
  sun_position,
  radiation_hourly,
  resolution = 0.1,
  item_pts = NULL,
  sequential = TRUE,
  xmin = -50,
  xmax = 50,
  ymin = -25,
  ymax = 50
)
}
\arguments{
\item{qsm}{An object of class \code{QSM}.}

\item{sun_position}{\code{data.frame}, sunlight direction over the time.}

\item{radiation_hourly}{\code{data.frame}, diffuse and global radiation over time.}

\item{resolution}{\code{numeric}, extent of the shading raster in
meters.}

\item{item_pts}{\code{matrix}, contains coordinates of simulated items.}

\item{sequential}{\code{boolean}, whether sequential (\code{TRUE}) or
parallel processing (\code{FALSE}) should be used.}

\item{xmin, xmax, ymin, ymax}{\code{numeric}, extent of the shading raster in
meters.}
}
\value{
\code{SpatRaster}, contains radiation around the tree for each
\code{sun_position}, the unit of the radiation depends on the unit used in
\code{radiation_hourly}. The single rasters for the time steps are stored as
layers with the timestamp as names.
}
\description{
\code{shade_tree}.
}
\details{
The parameter \code{sun_position} determines for which time steps the shade
is calculated. The \code{resolution} of the shade raster should correspond to
the unit used in \code{radiation_hourly}, e.g. when using a resolution of
0.1, the radiation should be given per dm². If the raster cell contains any
shade, the diffuse radiation is assigned. Otherwise, the global radiation is
assigned. Turning parallel processing on (\code{sequential = FALSE}) might
only work for windows computers.
}
\examples{
# load qsm
file_path <- system.file("extdata", "Prunus_avium_QSM_simplified.mat", package="qsm2shade")
qsm <- qsm2r::readQSM(file_path)

# get sun position at different times
timeframe <- seq(ISOdate(2020, 03, 22, 0, 0), ISOdate(2020, 03, 22, 23, 50), "10 mins")
sun_position <- sun_movement(timeframe, latitude = 48.07, longitude = 7.60)

# create dummy radiation data
radiation_hourly <- dummy_radiation_hourly(ISOdate(2020, 01, 01, 0, 0), ISOdate(2020, 12, 31, 23, 50))

# without leaves:

# calculate shade
result <- shade_tree(qsm, sun_position, radiation_hourly)

# show shade
plot(result)

# with leaves:

# create polygons for single item
leaf <- create_single_leaf(leaf_type = "normal", length_m = 0.1)

# get dummy item distribution
distribution <- dummy_item_distribution()

# create items
leaf_pts <- add_items(qsm, distribution, leaf, item_type = "leaves")

# calculate shade
result <- shade_tree(qsm, sun_position, radiation_hourly, item_pts = leaf_pts)

# show shade
plot(result)
}
\seealso{
\code{\link{shade_summarize}}, \code{\link{sun_movement}},
\code{\link{add_items}}
}
