% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shading.R
\name{shade_tree_geoms}
\alias{shade_tree_geoms}
\title{Calculate shade of geoms}
\usage{
shade_tree_geoms(
  geom_wood,
  sun_position,
  geom_other = NULL,
  empty_grid = terra::rast(nlyrs = 1, res = 0.1, vals = 0, xmin = min(geom_wood[, 2]) -
    20, xmax = max(geom_wood[, 2]) + 20, ymin = min(geom_wood[, 3]) - 20, ymax =
    max(geom_wood[, 3]) + 20),
  plane_origin = c(median(geom_wood[, 2]), median(geom_wood[, 3]), min(geom_wood[, 4])),
  plane_normal = c(0, 0, 1),
  transparency = 0,
  sequential = TRUE
)
}
\arguments{
\item{geom_wood}{\code{matrix}, contains matrix with IDs and coordinates of
wood geoms.}

\item{sun_position}{\code{data.frame}, sunlight direction over the time.}

\item{geom_other}{\code{matrix}, contains matrix with IDs and coordinates of
leaf / flower geoms.}

\item{empty_grid}{\code{SpatRaster}, with target shade raster location and
spatial resolution.}

\item{plane_origin}{\code{numeric}, \code{xyz}-vector of a point on the
ground.}

\item{plane_normal}{\code{numeric}, \code{xyz}-vector of the ground normal.}

\item{transparency}{\code{numeric}, transparency of the geom shade,
1 = fully transparent, 0 = fully opaque.}

\item{sequential}{\code{boolean}, whether sequential (\code{TRUE}) or
parallel processing (\code{FALSE}) should be used.}
}
\value{
\code{SpatRaster}, contains light / shade / in-between around the tree for
each \code{sun_position}. The single rasters for the time steps are stored as
layers with the timestamp as names.
}
\description{
\code{shade_tree_geoms} calculates the shade cast by a tree based on given sun
positions. The shade is coded as 0 and light is coded as 1. The function can
include the shade of modeled items (leaves, flowers) if provided. The
processing may be done sequentially or in parallel. An empty output raster
with the desired extent and resolution is required. Further, the transparency
of leaves may be configured. The ground can be specified via a point on the
plane and the plane normal. Per default, an even ground at the origin is
assumed.
}
\details{
The parameter \code{sun_position} determines for which time steps the shade
is calculated.
}
\examples{
# load wood geoms
file_path <- system.file("extdata", "pear_wood.txt", package="qsm2shade")
geom_wood <- as.matrix(read.table(file_path, header = TRUE))

# load leaf geoms
file_path <- system.file("extdata", "pear_leaves.txt", package="qsm2shade")
geom_other <- as.matrix(read.table(file_path, header = TRUE))

# get sun position at different times
timeframe <- seq(ISOdate(2020, 03, 22, 0, 0), ISOdate(2020, 03, 22, 23, 50), "10 mins")
sun_position <- sun_movement(timeframe, latitude = 48.07, longitude = 7.60)

# create dummy radiation data
radiation <- dummy_radiation(
  ISOdate(2020, 01, 01, 0, 0), ISOdate(2020, 12, 31, 23, 50), "1 hour")

# calculate shade
result <- shade_tree_geoms(
  geom_wood, sun_position = sun_position, geom_other = geom_other,
  transparency = 0.7)

# create dummy radiation data
radiation <- dummy_radiation(
  ISOdate(2020, 01, 01, 0, 0), ISOdate(2020, 12, 31, 23, 50), "1 hour")

# add radiation data
result <- add_radiation(result, radiation)

# summarize per day
result_daily <- shade_summarize(result, "day", "sum")

# plot daily shade
terra::plot(result_daily)
}
\seealso{
\code{\link{shade_summarize}}, \code{\link{sun_movement}}, \code{\link{add_radiation}}
}
