% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shading.R
\name{sun_movement}
\alias{sun_movement}
\title{Calculate sunlight position over time}
\usage{
sun_movement(timeframe, latitude, longitude, timezone = 0)
}
\arguments{
\item{timeframe}{\code{POSIXct}, vector with times.}

\item{latitude}{\code{numeric}, latitude of the tree position.}

\item{longitude}{\code{numeric}, longitude of the tree position.}

\item{timezone}{\code{integer}, timezone of time data.}
}
\value{
\code{data.frame}, contains sunlight direction over the given time.
}
\description{
\code{sun_movement} calculate the direction of the sunlight over time to be
used by \code{shade_tree_qsm()} or \code{shade_tree_geoms()}..
}
\examples{
# get sun position at different times
timeframe <- seq(ISOdate(2020, 03, 22, 0, 0), ISOdate(2020, 03, 22, 23, 50), "10 mins")
sun_position <- sun_movement(timeframe, latitude = 48.07, longitude = 7.60)

# display values
head(sun_position)
}
\seealso{
\code{\link{shade_tree_qsm}}
}
