% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulating.R
\name{add_items}
\alias{add_items}
\title{Add leaves or flowers to QSM}
\usage{
add_items(
  qsm,
  item_distribution,
  item_poly,
  stem_len = 0.01,
  item_type = c("leaves", "flowers"),
  item_angle = 45,
  cylinder_classes = NULL,
  add_noise = 0
)
}
\arguments{
\item{qsm}{An object of class \code{QSM}.}

\item{item_distribution}{\code{data.frame}, describes the item distribution,
with columns \code{c("diam_start_m", "diam_end_m", "m_per_item",
"item_scaling")}, optionally including variables from \code{classify_crown()}.}

\item{item_poly}{\code{matrix} or \code{list}, item to be simulated,
contains coordinates of a single polygon in a matrix or a list of matrices.}

\item{stem_len}{\code{numeric}, item stem length.}

\item{item_type}{\code{character}, type of added items, either
\code{"leaves"} or \code{"flowers"}}

\item{item_angle}{\code{numeric}, the angle between the ground and the leaf
axis in degrees when simulating leaves.}

\item{cylinder_classes}{\code{matrix} (optional), crown classification of
each cylinder when using an item distribution including variables from
\code{classify_crown()}.}

\item{add_noise}{\code{numeric} (optional), fraction of the stem length used
for adding noise to the item location and stem length, 0 means no noise.}
}
\value{
\code{matrix}, contains coordinates of the simulated items. For each polygon
node of the initial \code{item_poly}, three columns for the xyz-coordinates
are given.
}
\description{
\code{add_items} simulates leaves or flowers for a \code{QSM} object.
}
\examples{
# load qsm
file_path <- system.file("extdata", "Prunus_avium_QSM_simplified.mat", package="qsm2shade")
qsm <- qsm2r::readQSM(file_path)

# shift qsm to origin
# (shade is always projected to z = 0)
qsm <- qsm2r::set_location(qsm, c(0,0,0))

# create polygons for single item
flower <- create_single_flower()

# without classification:

# create dummy item regression
distribution <- dummy_item_distribution()

# add items
add_items(qsm, distribution, flower, item_type = "flowers")

# with classification:

# classify crown cylinders
classes <- classify_crown(qsm, compass_directions = 8, outside_buffer_m = 2, vertical_sections = 3)

# create dummy item regression
distribution_classes <- dummy_item_distribution(classification = TRUE)

# add items
add_items(qsm, distribution_classes, flower, item_type = "flowers", cylinder_classes = classes)
}
\seealso{
\code{\link{dummy_item_distribution}},
\code{\link{create_single_flower}}, \code{\link{create_single_leaf}}
}
