% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifying.R
\name{classify_crown}
\alias{classify_crown}
\title{Classify crown cylinders into crown sections}
\usage{
classify_crown(
  qsm,
  crs_epsg = 25832,
  compass_directions = 8,
  min_cylinders_cbh = 3,
  vertical_sections = 3,
  outside_buffer_m = 2
)
}
\arguments{
\item{qsm}{An object of class \code{QSM}.}

\item{crs_epsg}{\code{integer}, EPSG code of the coordinate reference system.}

\item{compass_directions}{\code{integer}, number of compass directions.}

\item{min_cylinders_cbh}{\code{integer}, minimum number of cylinders of the lowest branch
defining crown base height.}

\item{vertical_sections}{\code{integer}, number of vertical crown sections.}

\item{outside_buffer_m}{\code{numeric}, width of the buffer defining the outer and inner
horizontal section in meters.}
}
\value{
\code{matrix}, contains three columns: \code{compass} (compass directions),
\code{vertical} (vertical crown sections) and \code{horizontal} (horizontal
crown sections).
}
\description{
\code{classify_crown} classifies the cylinders of a \code{QSM} object into
separate  crown sections. It can use a variable number of compass directions
and vertical sections and classify the crown into an outer and an inner
horizontal section.
}
\examples{
# load qsm
file_path <- system.file("extdata", "Prunus_avium_QSM_simplified.mat", package="qsm2shade")
qsm <- qsm2r::readQSM(file_path)

# classify crown cylinders
classify_crown(qsm)
}
\seealso{
\code{\link{add_items}}
}
