% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulating.R
\name{create_flower}
\alias{create_flower}
\title{Create single flower polygons}
\usage{
create_flower(radius_m = 0.01)
}
\arguments{
\item{radius_m}{\code{numeric}, radius of the pentagon for the flower base
in meters.}
}
\value{
\code{matrix}, contains coordinates of a single flower.
}
\description{
\code{create_flower} creates polygons of a single flower to be used
by \code{add_geoms()}.
}
\examples{
# create polygons for single flower
flower <- create_flower()
}
\seealso{
\code{\link{add_geoms}}, \code{\link{plot_geom}}
}
