% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulating.R
\name{create_leaf}
\alias{create_leaf}
\title{Create single leaf polygon}
\usage{
create_leaf(
  type = c("normal", "ginkgo", "walnut", "3d_up", "3d_down", "3d_acer"),
  length_m = 0.01
)
}
\arguments{
\item{type}{\code{character}, shape of the leaf, should be one of
\code{c("normal", "ginkgo", "walnut", "3d_up", "3d_down", "3d_acer")}.}

\item{length_m}{\code{numeric}, length of a single leaf in meters.}
}
\value{
\code{matrix}, contains coordinates of a single leaf.
}
\description{
\code{create_leaf} creates a polygon of a single leaf to be used
by \code{add_geoms()}.
}
\examples{
# create polygon for single leaf
leaf <- create_leaf("3d_acer")

# display values
print(leaf)

# plot
plot_geoms(leaf, add = FALSE, axes = TRUE)
}
\seealso{
\code{\link{add_geoms}}, \code{\link{plot_geom}}
}
