% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulating.R
\name{create_leaf}
\alias{create_leaf}
\title{Create single leaf polygon}
\usage{
create_leaf(type = c("normal", "heart", "feather"), length_m = 0.01)
}
\arguments{
\item{type}{\code{character}, shape of the leaf, should be one of
\code{c("normal", "heart", "feather")}.}

\item{length_m}{\code{numeric}, length of a single leaf in meters.}
}
\value{
\code{matrix}, contains coordinates of a single leaf.
}
\description{
\code{create_leaf} creates a polygon of a single leaf to be used
by \code{add_geoms()}.
}
\examples{
# create polygon for single leaf
leaf <- create_leaf()
}
\seealso{
\code{\link{add_geoms}}, \code{\link{plot_geom}}
}
