% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geom_tree_location}
\alias{geom_tree_location}
\title{Extract stem coordinates from wood geoms}
\usage{
geom_tree_location(geom, lwr_height = 0.3, upr_height = 0.6)
}
\arguments{
\item{geom}{\code{matrix}, contains matrix with IDs and coordinates of wood
geoms.}

\item{lwr_height}{\code{numeric}, lower height threshold in meters.}

\item{upr_height}{\code{numeric}, upper height threshold in meters.}
}
\value{
A \code{numeric} containing the \code{xyz}-coordinates of the stem center.
}
\description{
\code{geom_tree_location} extracts the coordinates of the stem of a single
tree at a specified height range. Assumes that the input geoms describe
only one tree. The heights refer to the height over the ground.
}
\examples{
# load wood geoms
file_path <- system.file("extdata", "pear_wood.txt", package="qsm2shade")
geom_wood <- read.table(file_path, header = T

# get stem location
location <- las_tree_location(geom_wood)
}
\seealso{
\code{\link{plot_ground}}
}
