% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_geoms}
\alias{plot_geoms}
\title{Plot polygons}
\usage{
plot_geoms(geoms, col = "#6D9DC5", add = TRUE, lit = TRUE)
}
\arguments{
\item{geoms}{\code{matrix}, contains coordinates of simulated leaves /
flowers.}

\item{col}{\code{character}, color of the geoms.}

\item{add}{\code{boolean}, add the plot to current active \code{rgl} plot.}

\item{lit}{\code{boolean}, whether the polygons should be lit.}
}
\value{
\code{rgl} plot of all geoms.
}
\description{
\code{plot_geoms} plots geoms, e.g. those simulated by \code{add_geoms()}.
}
\examples{
# load qsm
file_path <- system.file("extdata", "walnut.mat", package="qsm2shade")
qsm <- qsm2r::readQSM(file_path)

# create dummy geom regression
distribution <- dummy_geom_distribution()

# create polygons for single geom
leaf <- create_leaf(type = "normal", length_m = 0.1)

# add geoms
leaves <- add_geoms(qsm, distribution, leaf, geom_type = "leaf")

# plot qsm
qsm2r::plot(qsm, col = "salmon4", lit = TRUE)

# plot geoms
plot_geoms(leaves, col = "darkolivegreen3")

# or:

# create polygons for single geom
flower <- create_flower(radius_m = 0.02)

# add geoms
flowers <- add_geoms(qsm, distribution, flower, geom_type = "flower")

# plot qsm
qsm2r::plot(qsm, col = "salmon4", lit = TRUE)

# plot geoms
plot_geoms(flowers, col = "steelblue3")

# or:

#' # load wood polygons
file_path <- system.file("extdata", "pear_wood.txt", package="qsm2shade")
poly_wood <- read.table(file_path, header = T)

# load leaf polygons
file_path <- system.file("extdata", "pear_leaves.txt", package="qsm2shade")
poly_leaves <- read.table(file_path, header = T)

# plot wood & leaves (takes quite some time)
plot_geoms(poly_wood, col = "salmon4", add = F)
plot_geoms(poly_leaves, col = "darkolivegreen3", add = T)
}
\seealso{
\code{\link{add_geoms}}, \code{\link{plot_shade_geoms}}
}
