% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_ground}
\alias{plot_ground}
\title{Plot ground}
\usage{
plot_ground(
  plane_origin = c(0, 0, 0),
  plane_normal = c(0, 0, 1),
  radius = 12,
  n_dir = 30L,
  z_offset = -0.005,
  col = "#B6CC8F",
  add = TRUE,
  lit = FALSE
)
}
\arguments{
\item{plane_origin}{\code{numeric}, \code{xyz}-vector of a point on the
ground.}

\item{plane_normal}{\code{numeric}, \code{xyz}-vector of the ground normal.}

\item{radius}{\code{numeric}, circle radius in meters.}

\item{n_dir}{\code{integer}, the number of directions at which points on the
circle are calculated.}

\item{z_offset}{\code{numeric}, offset along the z-axis in meters.}

\item{col}{\code{character}, color of the shade.}

\item{add}{\code{boolean}, add the plot to current active \code{rgl} plot.}
}
\value{
\code{rgl} plot of the ground.
}
\description{
\code{plot_ground} plots a circle of ground. The ground is specified via a
point on the plane and the plane normal. For plotting, the radius of the
circle and the number of points in the circle outline is required. To
prevent plotted shade of being obscured by the ground, the ground is shifted
by a small offset along the z-axis.
}
\examples{
# load qsm
file_path <- system.file("extdata", "walnut.mat", package="qsm2shade")
qsm <- qsm2r::readQSM(file_path)

# shift qsm to origin
# (shade is always projected to z = 0)
qsm <- qsm2r::set_location(qsm, c(0,0,0))

# plot qsm
qsm2r::plot(qsm, col = "salmon4", lit = TRUE)

# define ground
ground_origin <- c(0,0,0)
ground_normal <- c(0.1,0.05,1)

# plot ground
plot_ground(plane_origin = ground_origin, plane_normal = ground_normal, radius = 8)

# plot shade of wood
plot_shade_qsm(qsm, plane_origin = ground_origin, plane_normal = ground_normal)
}
\seealso{
\code{\link{plot_shade}}, \code{\link{plot_shade_geoms}}
}
