% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_shade_wood}
\alias{plot_shade_wood}
\title{Plot shadows cast by wood}
\usage{
plot_shade_wood(
  qsm,
  sun_direction = c(0.25, 0.5, -0.75),
  plane_origin = c(0, 0, 0),
  plane_norm = c(0, 0, 1),
  col = "grey40",
  add = TRUE
)
}
\arguments{
\item{qsm}{An object of class \code{QSM}.}

\item{sun_direction}{\code{numeric}, vector containing the unit vector of the
sun direction.}

\item{plane_origin}{\code{numeric}, \code{xyz}-vector of a point on the
ground.}

\item{plane_norm}{\code{numeric}, \code{xyz}-vector of the ground normal.}

\item{col}{\code{character}, color of the shade.}

\item{add}{\code{boolean}, add the plot to current active \code{rgl} plot.}
}
\value{
\code{rgl} plot of the shade.
}
\description{
\code{plot_shade_wood} plots shadows of a \code{QSM} object. The ground can
be specified via a point on the plane and the plane normal. Per default, an
even ground at the coordinate origin is assumed.
}
\examples{
# load qsm
file_path <- system.file("extdata", "walnut.mat", package="qsm2shade")
qsm <- qsm2r::readQSM(file_path)

# shift qsm to origin
# (shade is always projected to z = 0)
qsm <- qsm2r::set_location(qsm, c(0,0,0))

# plot qsm
qsm2r::plot(qsm, col = "salmon4", lit = TRUE)

# plot shade of wood
plot_shade_wood(qsm)
}
\seealso{
\code{\link{plot_shade_items}}
}
