% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shading.R
\name{shade_summarize}
\alias{shade_summarize}
\title{Summarize shade per period}
\usage{
shade_summarize(radiation_grid, period = c("day", "month", "year", "total"))
}
\arguments{
\item{radiation_grid}{\code{SpatRaster}, stacked set of shade rasters from
\code{shade_tree()}.}

\item{period}{\code{character}, length of summary period, either
\code{"day"}, \code{"month"}, \code{"year"} or \code{"total"}.}
}
\value{
\code{SpatRaster}, stacked shade rasters summarized by time periods, the
layer names indicate the respective time period.
}
\description{
\code{shade_summarize} summarizes the rasters produced by \code{shade_tree()}
in variable time periods.
}
\examples{
# load qsm
file_path <- system.file("extdata", "Prunus_avium_QSM_simplified.mat", package="qsm2shade")
qsm <- qsm2r::readQSM(file_path)

# get son position at different times
sun_position <- sun_movement(ISOdate(2020, 03, 22, 0, 0), ISOdate(2020, 03, 23, 23, 50), "10 mins", lat = 48.07, lon = 7.60)

# create dummy radiation data
radiation_hourly <- dummy_radiation_hourly(ISOdate(2020, 01, 01, 0, 0), ISOdate(2020, 12, 31, 23, 50))

# calculate shade
result <- shade_tree(qsm, sun_position, radiation_hourly)

# summarize per day
result_daily <- shade_summarize(result, "day")

# plot daily shade
terra::plot(result_daily)
}
\seealso{
\code{\link{shade_tree}}
}
